requirement.txt
Django==2.0
django-bootstrap3==11.0.0
pytz==2018.5
tencentcloud-sdk-python==3.0.22
